Fix registry key, remove invisible unicode chars
One of the registry keys related to the Logon Script
Environment persistence mechanism incorrectly listed
'UserMprLogonScript' when it should be
'UserInitMprLogonScript' (per the reference materials.)
Also, two of the registry keys had the invisible soft
hyphen character inserted.  I'm not sure if this affects
things when these data files are used for detection,
but it's probably best to remove them.